requirements.txt
python-dotenv==0.19.0
PyInquirer==1.0.3